front change

+++ src/Web/cube/src/components/boardAll.vue
@@ -1,9 +1,5 @@
 ﻿<template>
     <div style="margin:0">
-
-        <!--<div style="text-align:right;margin:10px;">
-        <Button type="primary" @click="AddBoard()">AddBoard</Button>
-    </div>-->
         <ul>
             <li style="width:260px; float: left;">
                 <Card style="width: 250px; cursor: pointer;" v-on:click.native="AddBoard()">
@@ -23,12 +19,12 @@
                         <img :src="getUserAvatar(board.CreatedUser)" :title="board.CreatedUser" style="width:20px; height:20px; border-radius:50%; " />
                     </p>
                     <div style="text-align:center;">
-                        <a href="#" slot="extra" @click.prevent="ViewBoard(board.Id)">
+                        <a href="#" slot="extra" @click.prevent="ViewBoard(board)">
                             <Icon type="ios-loop-strong"></Icon>
                             {{board.Name}}
                         </a>
                     </div>
-                    <a href="#" slot="extra" @click.prevent="DeleteBoard(board)" title="Delete">
+                    <a href="#" slot="extra" @click.prevent="DeleteBoard(board)" title="Delete" v-if="board.CreatedUser==UserName">
                         <span aria-label="Delete" class="">
                             <button class="css-b7766g" tabindex="-1" type="button" aria-label="Delete" style="position: relative; padding-left: 0px; padding-right: 0px; min-width: 42px;">
                                 <svg class="css-vubbuv" focusable="false" viewBox="0 0 24 24" aria-hidden="true" style="color: rgb(239, 83, 80);">
@@ -52,7 +48,7 @@
             return {
                 UserToken: null,
                 post: null,
-
+                UserName:null,
                 connection: "",
             };
         },
@@ -61,6 +57,8 @@
             // already being observed
             this.fetchData();
             this.UserToken = localStorage.getItem('TOKEN');
+            this.UserName = localStorage.getItem('LOGINUSER');
+            
             this.init();
         },
         methods: {
@@ -68,6 +66,7 @@
                 fetch('Board')
                     .then(r => r.json())
                     .then(json => {
+                        console.log(json);
                         this.post = json;
                         return;
                     });
@@ -75,22 +74,35 @@
             AddBoard() {
                 this.$router.push('/addboard');
             },
-            ViewBoard(val) {
-                this.$router.push({ name:'boardDetail', params: { boardId: val } });
+            ViewBoard(board) {
+                this.$router.push({ name: 'boardDetail', params: { boardId: board.Id, boardName: board.Name, state:board.State } });
             },
             DeleteBoard(board) {
-                this.axios.delete('/Board/' + board.Id + '',
+                this.$confirm(
                     {
-                        headers: {
-                            'Authorization': 'Bearer ' + this.UserToken
+                        message: 'Are you sure delete board [' + board.Name + '] ?',
+                        button: {
+                            no: 'No',
+                            yes: 'Yes'
+                        },
+                        callback: confirm => {
+                            if (confirm) {
+                                this.axios.delete('/Board/' + board.Id + '',
+                                    {
+                                        headers: {
+                                            'Authorization': 'Bearer ' + this.UserToken
+                                        }
+                                    }).then(() => {
+                                        this.fetchData();
+                                    }).then(() => {
+                                        this.renderFunc(board.Name + ' is deleted successfully.');
+                                    }).then(() => {
+                                        this.sendMsg();
+                                    })
+                            }
                         }
-                    }).then(() => {
-                        this.fetchData();
-                    }).then(() => {
-                        this.renderFunc(board.Name + ' is deleted successfully.');
-                    }).then(() => {
-                        this.sendMsg();
-                    })
+                    }
+                )
             },
             getUserAvatar(userName) {
                 let userAvatar = ""
